Extract Msf::DBManager::Import
MSP-11124

Extract all methods dealing with imports.
